Our docs say we use strict percent encoding but we don't "really"
Specifically, we use strict percent encoding during the signing
step, but when we return URLs, we use `urllib.urlencode`, which
is not the same. This patch is rather 'fugly', but enforces
strict percent encoding for URLs returned by the client as well.

This needs some discussion about whether it's even necessary...
but here is a PoC.