Fix collections.Iterable import on Python 3.8 (#119)
In Python 3.3 collections' Abstract Base Classes (ABCs) moved to the
`collections.abc` namespace. A compatibility shim made it possible to
still import them from the old location, but in Python 3.7 you get a
DeprecationWarning, and in Python 3.8 it will simply not work.

This change assumes we're running on Python 3.3 or above, but falls back
to import from the old location if that assumption fails.